requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
confluent-kafka>=2.12.2
google-cloud-aiplatform==1.132.0
google-cloud-secret-manager==2.26.0
ddtrace==4.1.1
requests==2.32.4
python-dotenv==1.2.1
pydantic==2.12.5